refactor: channel-element-hash -> icon
