+++ src/components/Main/NavigationBar/ChannelList/ChannelElementIcon.vue
